+++ src/Journalist.EventStore/Events/JournaledEventPropertyNames.cs
@@ -5,12 +5,14 @@
         public static readonly string EventId = "EventId";
         public static readonly string EventType = "EventType";
         public static readonly string EventPayload = "EventPayload";
+        public static readonly string EventHeaders = "EventHeaders";
 
         public static readonly string[] All =
         {
             EventId,
             EventType,
-            EventPayload
+            EventPayload,
+            EventHeaders
         };
     }
 }